requirements.txt
 --extra-index-url https://download.pytorch.org/whl/cu117
 
absl-py==2.1.0
accelerate==0.29.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==4.3.0
anykeystore==0.2
apex==0.9.10.dev0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cmake==3.29.0.1
codecov==2.1.13
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
coverage==7.4.4
cryptacular==1.6.2
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.19.1
dearpygui==1.8.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.18.2
dill==0.3.8
einops==0.6.0
exceptiongroup==1.2.0
executing==2.0.1
fairscale==0.4.13
fast_simplification==0.1.7
fastapi==0.110.1
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.3
flake8==7.0.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.0
future==1.0.0
fuzzywuzzy==0.18.0
gdown==5.1.0
gradio==4.18.0
gradio_client==0.10.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
hupper==1.12.1
idna==3.6
image-reward==1.5
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
iniconfig==2.0.0
interrogate==1.7.0
ipykernel==6.29.4
ipython==8.23.0
ipywidgets==8.1.2
isoduration==20.11.0
isort==4.3.21
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.4
Levenshtein==0.25.1
lightning-utilities==0.11.2
lit==18.1.2
llvmlite==0.42.0
lpips==0.1.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
MiniballCpp==0.2.3
mistune==3.0.2
mmcls==0.25.0
mmcv @ git+https://github.com/open-mmlab/mmcv.git@d409eedc816fccfb1c8d57e5eed5f03bd075f327
mmgen @ git+https://github.com/Lakonik/mmgeneration.git@500a93e474638c87dcaf8fad94cfbe1ef29acd1f
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.17.3
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.0
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.4
PasteDeploy==3.1.0
pbkdf2==1.3
pexpect==4.9.0
pillow==10.3.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
prettytable==3.10.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==5.26.1
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyflakes==3.2.0
pygltflib==1.16.1
Pygments==2.17.2
PyMatting==1.1.12
PyMCubes==0.1.2
pyparsing==3.1.2
pyramid==2.0.2
pyramid-mailer==0.15.1
PySocks==1.7.1
pytest==8.1.1
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-Levenshtein==0.25.1
python-multipart==0.0.9
python3-openid==3.2.0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
rapidfuzz==3.9.1
referencing==0.34.0
regex==2023.12.25
rembg==2.0.56
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.5
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.13.0
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
semantic-version==2.10.0
Send2Trash==1.8.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.1
tifffile==2024.2.12
timm==0.6.13
tinycss2==1.3.0
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.0.1
torch.redstone==0.0.5
torchmetrics==1.3.2
torchsde==0.2.6
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
trampoline==0.1.2
transaction==4.0
transformers==4.39.3
translationstring==1.4
trimesh==4.1.4
triton==2.0.0
typer==0.12.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.29.0
velruse==1.1.1
venusian==3.1.0
videoio==0.2.8
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
WebOb==1.8.7
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.1.2
wtforms-recaptcha==0.3.2
xatlas==0.0.9
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
zope.deprecation==5.0
zope.interface==6.2
zope.sqlalchemy==3.1